+++ include/CLI/REPL/Repl.hpp
@@ -2,24 +2,25 @@
  * @file Repl.hpp
  * @author Alexandre Plateau (lexplt.dev@gmail.com)
  * @brief ArkScript REPL - Read Eval Print Loop
- * @version 0.2
+ * @version 1.0
  * @date 2020-10-27
  *
- * @copyright Copyright (c) 2020-2021
+ * @copyright Copyright (c) 2020-2024
  *
  */
 
 #ifndef ARK_REPL_REPL_HPP
 #define ARK_REPL_REPL_HPP
 
-#include <iostream>
-#include <filesystem>
+#include <string>
+#include <optional>
 
 #include <Ark/Constants.hpp>
 #include <Ark/Compiler/Compiler.hpp>
 #include <Ark/VM/VM.hpp>
 #include <Ark/VM/State.hpp>
-#include <CLI/REPL/ConsoleStyle.hpp>
+
+#include <replxx.hxx>
 
 namespace Ark
 {
@@ -29,9 +30,9 @@ namespace Ark
         /**
          * @brief Construct a new Repl object
          *
-         * @param libenv search path for the std library
+         * @param lib_env search path for the std library
          */
-        explicit Repl(const std::vector<std::filesystem::path>& libenv);
+        explicit Repl(const std::vector<std::filesystem::path>& lib_env);
 
         /**
          * @brief Start the REPL
@@ -40,16 +41,29 @@ namespace Ark
         int run();
 
     private:
-        Replxx m_repl;
-        unsigned m_lines;
-        int m_old_ip;
-        std::vector<std::filesystem::path> m_libenv;
+        replxx::Replxx m_repl;
+        unsigned m_line_count;
+        std::string m_code;
+        bool m_running;
 
-        static inline void print_repl_header();
-        static int count_open_parentheses(const std::string& line);
-        static int count_open_braces(const std::string& line);
-        static void trim_whitespace(std::string& line);
-        void cui_setup();
+        int m_old_ip;
+        std::vector<std::filesystem::path> m_lib_env;
+        State m_state;
+        VM m_vm;
+        bool m_has_init_vm;
+
+        /**
+         * @brief Configure replxx
+         */
+        void cuiSetup();
+
+        /**
+         * @brief Get a line via replxx and handle commands
+         * @param continuation if the prompt needs to be modified because a code block isn't entirely closed, set to true
+         * @return
+         */
+        std::optional<std::string> getLine(bool continuation);
+        std::optional<std::string> getCodeBlock();
     };
 }