+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 - added boost-ext/ut to write unit tests in C++
 - basic ArkScript code formatter, available through the CLI: `arkscript -f|--format`
 - comments are now tracked in the AST and attached to the nearest node below them
+- `VM::forceReloadPlugins`, to be used by the REPL to force reload the plugins and be sure that their symbols are all define
+- added `help`, `save` (save history to disk), `history` (print history), `reset` (reset vm and code) commands to the REPL
+- REPL can now show when a code block isn't terminated (prompt changes from `>` to `:`)
+- more controls available inside the REPL
 
 ### Changed
 - instructions are on 4 bytes: 1 byte for the instruction, 1 byte of padding, 2 bytes for an immediate argument